move router filters into their lane
