+++ lib/kv-router/src/scheduling/mod.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod config;
+mod filter;
 mod local;
 pub mod overlap_refresh;
 pub mod policy;
@@ -10,6 +11,7 @@ pub mod queue;
 pub mod selector;
 
 mod types;
+pub use filter::*;
 pub use local::LocalScheduler;
 pub use overlap_refresh::{NoopOverlapScoresRefresh, OverlapScoresRefresh, RefreshedOverlap};
 pub use prefill_load::PrefillLoadEstimator;
